+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 ## Next version
 
+## 1.3.0
+
 - ApiControllerConcern : add `model_errors` in json response
 - `json_response` is now a `HashWithIndifferentAccess`